carved out token replacement into a seperate class (moved it out of TargetDir.cs), TokenReplacement is a static class because it's stateless and functional
