+++ warmup.Tests/describe_TokenReplacement.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using NSpec;
+
+namespace warmup.Tests
+{
+    class describe_TokenReplacement : nspec
+    {
+        void file_encodings()
+        {
+            it["preamble is retained after token replacement"] = () =>
+            {
+
+            };
+
+            it["preamble is added if the file didn't have one in the first place"] = () =>
+            {
+
+            };
+        }
+    }
+}
