+++ warmup/TargetDir.cs
@@ -65,7 +65,7 @@ namespace warmup
             MoveAllFiles(startingPoint, _replacementTokens);
 
             //replace file content
-            ReplaceTokensInTheFiles(startingPoint, _replacementTokens);
+            TokenReplacement.ReplaceTokensInTheFiles(startingPoint, _replacementTokens);
         }
 
         private void MoveAllDirectories(DirectoryInfo dir, IDictionary<string, string> tokens)
@@ -139,79 +139,6 @@ namespace warmup
                     }
                 }
             }
-        }
-
-        private void ReplaceTokensInTheFiles(DirectoryInfo point, IDictionary<string, string> tokens)
-        {
-            List<string> ignoredExtensions = GetIgnoredExtensions();
-
-            var fileContents = new StringBuilder();
-
-            foreach (var info in point.GetFiles("*.*", SearchOption.AllDirectories))
-            {
-                if (ignoredExtensions.Contains(info.Extension, StringComparer.InvariantCultureIgnoreCase)) continue;
-                //skip the .git directory
-                if (new[] {"\\.git\\"}.Contains(info.FullName)) continue;
-                // skip readonly and hidden files
-                if (info.IsReadOnly || (info.Attributes & FileAttributes.Hidden) == FileAttributes.Hidden) continue;
-
-
-                // process contents
-                fileContents.Clear();
-                fileContents.Append(File.ReadAllText(info.FullName));
-                //string contents = File.ReadAllText(info.FullName);
-                
-                // replace tokens
-                foreach (var token in tokens)
-                {
-                    fileContents.Replace(token.Key, token.Value);
-                }
-
-                var originalFileEncoding = GetFileEncoding(info.FullName);
-                File.WriteAllText(info.FullName, fileContents.ToString(), originalFileEncoding);
-            }
-        }
-
-        private Encoding GetFileEncoding(string fileName)
-        {
-            Encoding encoding = null;
-
-            var file = new FileStream(fileName, FileMode.Open, FileAccess.Read, FileShare.Read);
-            if (file.CanSeek)
-            {
-                //use ASCII as the default if it doesn't match any of the others.
-                encoding = Encoding.ASCII;
-
-                //determine encoding based on first 4 bytes of the file.
-                var bom = new byte[4];
-                file.Read(bom, 0, 4);
-                if (bom[0] == 0xef && bom[1] == 0xbb && bom[2] == 0xbf)
-                    encoding = Encoding.UTF8;
-                if (bom[0] == 0xff && bom[1] == 0xfe)
-                    encoding = Encoding.Unicode;
-                if (bom[0] == 0xfe && bom[1] == 0xff)
-                    encoding = Encoding.Unicode;
-                if (bom[0] == 0x00 && bom[1] == 0x00 && bom[2] == 0xfe && bom[3] == 0xff)
-                    encoding = Encoding.Unicode;
-
-                file.Close();
-            }
-            else
-            {
-                encoding = Encoding.ASCII;
-            }
-
-            return encoding;
-        }
-
-        private static List<string> GetIgnoredExtensions()
-        {
-            var extension = new List<string>();
-            foreach (IgnoredFileType ignoredFileType in WarmupConfiguration.settings.IgnoredFileTypeCollection)
-            {
-                extension.Add(string.Format(".{0}", ignoredFileType.Extension));
-            }
-            return extension;
         }
         
         public void MoveToDestination(string target)
